requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
itsdangerous
Jinja2